requirements.txt
Django==2.2.6
django-extensions==2.2.3
django-redis==4.10.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
